+++ app/src/main/java/com/example/gh_coursework/domain/entity/PointsTagsDomain.kt
@@ -1,6 +1,6 @@
 package com.example.gh_coursework.domain.entity
 
 data class PointsTagsDomain(
-    val pointId: Int,
+    val pointId: Long,
     val tagId: Int
 )
